Refactoring, NeedModel moved to new package named "model"
