+++ src/main/java/hu/drumbun/service/need/converter/NeedModelConverter.java
@@ -1,6 +1,6 @@
 package hu.drumbun.service.need.converter;
 
-import hu.drumbun.controller.need.NeedModel;
+import hu.drumbun.controller.need.model.NeedModel;
 import hu.drumbun.entities.Need;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
